+++ prisma/schema.prisma
@@ -79,8 +79,8 @@ model Invoices {
     clientName      String
     companyName     String
     invoiceItems    InvoiceItems[]
-    Organizations   Organizations  @relation(fields: [organizationsId], references: [id])
-    organizationsId String
+    Organizations   Organizations  @relation(fields: [organization_id], references: [id])
+    organization_id String
 
     // @@schema("public")
 }
