+++ src/app/dashboard/invoices/new/page.tsx
@@ -1,21 +1,16 @@
 import Layout from "~/components/Layout/Layout";
-import { Header, NoAsset } from "~/components/Assets";
-import { ReactElement } from "react";
 import { api } from "~/utils/api";
 import NewInvoiceForm from "~/components/invoices/newForm/newInvoiceForm";
 import { Toaster } from "~/components/ui/toaster";
 
-function Page() {
+export default function Page() {
   return (
+    <Layout>
     <main className="pl-5">
       <Toaster />
       <NewInvoiceForm />
     </main>
+    </Layout>
   );
 }
 
-Page.getLayout = function getLayout(page: ReactElement) {
-  return <Layout>{page}</Layout>;
-};
-
-export default Page;
