+++ src/app/dashboard/invoices/page.tsx
@@ -9,14 +9,13 @@ import InvoiceList from "~/components/invoices/invoiceTables/InvoiceList";
 import { Toaster } from "~/components/ui/toaster";
 
 
-function Page() {
-  const { data, isLoading } = api.invoices.all.useQuery({
-    organizationEmail: ""
-  });
+export default function Page() {
+  const { data, isLoading } = api.invoices.all.useQuery();
 
   return (
+    <Layout>
     <main className="pl-5">
-      <Header title="Your Invoices" href="/invoices/new" />
+      <Header title="Your Invoices" href="/dashboard/invoices/new" />
       <Toaster />
 
       {data?.length === 0 && (
@@ -24,17 +23,15 @@ function Page() {
           bigTitle="You haven't added any invoices yet"
           smallTitle="Add Invoices for your guests and start tracking your finances"
           c2a="Add Invoices"
-          c2aUrl="/invoices/new"
+          c2aUrl="/dashboard/invoices/new"
         />
       )}
       {data?.length !== 0 && !isLoading && <InvoiceList invoices={data!} />}
       {isLoading && <LoadingSkeleton />}
     </main>
+    </Layout>
   );
 }
 
-Page.getLayout = function getLayout(page: ReactElement) {
-  return <Layout>{page}</Layout>;
-};
 
-export default Page;
+
